refactor(scoreInfo): ScoreInfo 객체의 패키지를 domain에서 dto로 이동
객체의 역할을 생각했을 때 도메인 객체보다는 dto 역할에 더 가까우므로 패키지를 변경
