+++ src/main/java/lotto/dto/ScoreInfo.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.dto;
 
 import lotto.vo.Score;
